refactor!: switch to async-only client, split API into modular async endpoints, add built-in rate limiting, and update demos/tests for auth/no-content handling
