+++ README.md
@@ -1,7 +1,6 @@
 # energyid-py
 
-Python client for the [EnergyID API](https://api.energyid.eu).
-Synchronous, async, and pandas interfaces — covering records, meters, readings, groups, organizations, search, and more.
+Async Python client for the [EnergyID API](https://api.energyid.eu).
 
 ## Installation
 
@@ -12,21 +11,27 @@ pip install energyid-py
 ## Quick Start
 
 ```python
+import asyncio
 from energyid import JSONClient
 
-client = JSONClient(api_key="YOUR_API_KEY")
 
-member = client.get_member()
-records = client.get_member_records()
-record = records[0]
+async def main():
+    async with JSONClient(api_key="YOUR_API_KEY") as client:
+        member = await client.get_member()
+        records = await client.get_member_records()
+        meters = await client.get_record_meters(records[0].id)
+        readings = await client.get_meter_readings(meters[0].id, take=10)
+        print(member["fullName"], len(readings["readings"]))
 
-meters = record.get_meters()
-readings = meters[0].get_readings(take=10)
+
+asyncio.run(main())
 ```
 
-### Authentication
+## Authentication
 
 ```python
+from energyid import JSONClient
+
 # API Key (recommended)
 client = JSONClient(api_key="YOUR_API_KEY")
 
@@ -35,129 +40,62 @@ client = JSONClient(device_token="YOUR_DEVICE_TOKEN")
 
 # OAuth2
 client = JSONClient(
-    client_id="ID", client_secret="SECRET",
-    username="user@example.com", password="pass",
+    client_id="ID",
+    client_secret="SECRET",
+    username="user@example.com",
+    password="pass",
 )
 ```
 
-## Features
+## Built-in Request Throttling
 
-### Rich Model Objects
+`JSONClient` includes two safeguards to reduce backend pressure:
 
-All API responses are wrapped in dict-subclass models (`Member`, `Record`, `Meter`, `Group`, `Organization`) that provide convenience methods while remaining fully dict-compatible:
+- `max_concurrency`: max in-flight HTTP requests.
+- `max_requests_per_window` + `rate_limit_window_seconds`: request-rate cap.
 
-```python
-record = client.get_record(record_id=123)
-record["displayName"]       # dict access
-record.id                   # property shortcut
-record.timezone             # "Europe/Brussels"
-
-meters = record.get_meters()
-meter = meters[0]
-meter.get_readings(take=5)
-meter.get_latest_reading()
-meter.create_reading("2024-06-01", 12345)
-```
-
-### Records
+Defaults are conservative (`max_concurrency=10`, `max_requests_per_window=20`, `rate_limit_window_seconds=1.0`).
 
 ```python
-# Data & definitions
-definitions = record.get_definitions()
-data = record.get_data("electricityImport", "2024-01-01", "2024-12-31", interval="month")
-benchmark = record.get_benchmark("electricityImport", year=2024)
-
-# Activity, timeline, directives, limits
-activity = record.get_activity(take=10)
-timeline = record.get_timeline("2020-01-01", "2025-12-31")
-directives = record.get_directives()
-limits = record.get_limits()
-
-# CRUD
-new_record = client.create_record("My Home", "household", "Brussels", "1000", "bE", "dwelling")
-record.edit(displayName="Updated Name")
-record.delete()
+client = JSONClient(
+    api_key="YOUR_API_KEY",
+    max_concurrency=5,
+    max_requests_per_window=10,
+    rate_limit_window_seconds=1.0,
+)
 ```
 
-### Meters
-
-```python
-meter = record.create_meter("Solar", "electricityExport", "kilowattHour", "counter")
-meter.create_reading("2024-06-01", 12345)
-meter.edit(displayName="Renamed")
-meter.hide()
-meter.close()
-meter.delete()
-```
-
-### Groups & Organizations
-
-```python
-groups = record.get_groups()
-group = client.get_group(groups[0].id)
-
-for member in group.get_members(amount=100):
-    print(member["fullName"])
-
-for rec in group.get_records(amount=50):
-    print(rec["displayName"])
-
-org = client.get_organization("org-id")
-org.get_groups()
-```
-
-### Search
-
-```python
-client.search_groups(q="solar")
-client.search_services(q="Fluvius", top=5)
-client.search_cities("bE", "Gent")
-client.get_meter_catalog()
-```
-
-### Transfers
-
-```python
-client.create_transfer("EA-123", "newowner@example.com")
-client.accept_transfer("transfer-id")
-client.decline_transfer("transfer-id")
-```
+Set `max_concurrency=None` and/or `max_requests_per_window=None` to disable a limiter.
 
 ## PandasClient
 
-Returns pandas DataFrames and Series instead of raw dicts:
+Use `PandasClient` for DataFrame/Series output:
 
 ```python
+import asyncio
 from energyid import PandasClient
 
-pc = PandasClient(api_key="YOUR_API_KEY")
 
-df = pc.get_meter_readings(meter_id="meter-id", take=100)
-ts = pc.get_record_data(record_id=123, name="electricityImport",
-                        start="2024-01-01", end="2024-12-31", interval="month")
+async def main():
+    async with PandasClient(api_key="YOUR_API_KEY") as client:
+        df = await client.get_meter_readings(meter_id="meter-id", take=100)
+        ts = await client.get_record_data(
+            record_id=123,
+            name="electricityImport",
+            start="2024-01-01",
+            end="2024-12-31",
+            interval="month",
+        )
+        print(df.shape, ts.shape)
+
+
+asyncio.run(main())
 ```
-
-## Async Client
-
-Full async mirror of all endpoints using `aiohttp`:
-
-```python
-from energyid.aio.client import JSONClient as AsyncJSONClient
-
-async with AsyncJSONClient(api_key="YOUR_API_KEY") as client:
-    member = await client.get_member()
-    records = await client.get_member_records()
-    meters = await client.get_record_meters(records[0].id)
-```
-
-## Demo
-
-See [Demo.ipynb](Demo.ipynb) for a complete walkthrough of all endpoints.
 
 ## API Documentation
 
 - API: https://api.energyid.eu/
-- Swagger spec: https://api.energyid.eu/api-reference/index.html
+- Swagger: https://api.energyid.eu/api-reference/index.html
 
 ## License
