+++ pyproject.toml
@@ -13,7 +13,6 @@ requires-python = ">=3.10"
 dependencies = [
     "aiohttp>=3.11.18",
     "pandas>=2.2.3",
-    "requests>=2.32.3",
 ]
 
 [build-system]
