+++ requirements.txt
@@ -1,0 +1,43 @@
+# This file was autogenerated by uv via the following command:
+#    uv export --no-hashes --output-file=requirements.txt --no-dev
+-e .
+aiohappyeyeballs==2.6.1
+    # via aiohttp
+aiohttp==3.11.18
+    # via energyid-py
+aiosignal==1.3.2
+    # via aiohttp
+async-timeout==5.0.1 ; python_full_version < '3.11'
+    # via aiohttp
+attrs==25.3.0
+    # via aiohttp
+frozenlist==1.6.0
+    # via
+    #   aiohttp
+    #   aiosignal
+idna==3.10
+    # via yarl
+multidict==6.4.3
+    # via
+    #   aiohttp
+    #   yarl
+numpy==2.2.5
+    # via pandas
+pandas==2.2.3
+    # via energyid-py
+propcache==0.3.1
+    # via
+    #   aiohttp
+    #   yarl
+python-dateutil==2.9.0.post0
+    # via pandas
+pytz==2025.2
+    # via pandas
+six==1.17.0
+    # via python-dateutil
+typing-extensions==4.13.2 ; python_full_version < '3.11'
+    # via multidict
+tzdata==2025.2
+    # via pandas
+yarl==1.20.0
+    # via aiohttp
