+++ src/energyid/aio/__init__.py
@@ -1,3 +1,5 @@
+from energyid.scope import Scope
+
 from .client import JSONClient, PandasClient
 
-__all__ = ["JSONClient", "PandasClient"]
+__all__ = ["JSONClient", "PandasClient", "Scope"]
