+++ src/energyid/aio/clients/endpoints/__init__.py
@@ -1,0 +1,19 @@
+from .catalogs import CatalogsMixin
+from .groups import GroupsMixin
+from .members import MembersMixin
+from .meters import MetersMixin
+from .organizations import OrganizationsMixin
+from .records import RecordsMixin
+from .search import SearchMixin
+from .transfers import TransfersMixin
+
+__all__ = [
+    "CatalogsMixin",
+    "GroupsMixin",
+    "MembersMixin",
+    "MetersMixin",
+    "OrganizationsMixin",
+    "RecordsMixin",
+    "SearchMixin",
+    "TransfersMixin",
+]
