+++ src/energyid/aio/clients/endpoints/catalogs.py
@@ -1,0 +1,3 @@
+class CatalogsMixin:
+    async def get_meter_catalog(self) -> dict:
+        return await self._request("GET", "catalogs/meters")
